Introduce importNodeIntoDocument()
