+++ model/index.js
@@ -45,6 +45,7 @@ export { default as Fragmenter } from './Fragmenter'
 export { default as getChangeFromDocument } from './getChangeFromDocument'
 export { default as HTMLExporter } from './HTMLExporter'
 export { default as HTMLImporter } from './HTMLImporter'
+export { default as importNodeIntoDocument } from './importNodeIntoDocument'
 export { default as IncrementalData } from './IncrementalData'
 export { default as InlineNode } from './InlineNode'
 export { default as JSONConverter } from './JSONConverter'
